+++ spring-blog/History.md
@@ -6,3 +6,4 @@ Steps of Implementation
 3. Create Repositories.
 4. [Setup Spring Security](Howto.md#1---configure-spring-security)
 5. [Implement Login using JWT](Howto.md#2---implement-login-using-jwt).
+6. [Validate JWT](Howto.md#3---validate-jwt).